Changes to the loop
`brightnessValue´, ´blackValue´, ´whiteValue´ have all become `modifier`. For better or worse.
A new function called `changeModifier` is in place for easy configuration of mid-loop stuff. The current configuration changes modifier for each iteration with a hint of randomness. If you want modifier to be static throughout, you can comment out line 71, or empty the function of its commands.